refactor: typeorm comment repo 파일 이름 변경
