+++ src/test/controller/comments.controller.spec.ts
@@ -2,7 +2,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { CommentsController } from '../../modules/comments/comments.controller';
 // import { CommentsService } from '../../modules/comments/comment.service';
 import { CommentsService } from '../../modules/comments/comments.service';
-import { CommentsRepository } from '../../modules/comments/comments.repository';
+import { CommentsRepository } from '../../modules/comments/comments-typeorm.repository';
 import { UsersRepository } from '../../modules/users/users-typeorm.repository';
 import { PostsRepository } from '../../modules/posts/posts-typeorm.repository';
