+++ src/test/repository/comments.repository.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { CustomTypeOrmModule } from '../../config/typeorm/custom-typeorm-module';
-import { CommentsRepository } from '../../modules/comments/comments.repository';
+import { CommentsRepository } from '../../modules/comments/comments-typeorm.repository';
 import { UsersRepository } from '../../modules/users/users-typeorm.repository';
 import { TestTypeOrmModule } from '../db/test-db.module';
 import { User } from '../../modules/users/entities/user.entity';
